+++ src/modules/edc-demo/edc-demo.module.ts
@@ -33,14 +33,25 @@ import {NgxJsonViewerModule} from 'ngx-json-viewer';
 import {AssetDetailDialog} from './components/asset-detail-dialog/asset-detail-dialog.component';
 import {AssetEditorDialog} from './components/asset-editor-dialog/asset-editor-dialog.component';
 import {AssetListComponent} from './components/asset-list/asset-list.component';
+import {AssetSelectComponent} from './components/asset-select/asset-select.component';
 import {AssetViewerComponent} from './components/asset-viewer/asset-viewer.component';
-import {CatalogBrowserFetchDetailDialogComponent} from './components/catalog-browser-fetch-detail-dialog/catalog-browser-fetch-detail-dialog.component';
+import {
+  CatalogBrowserFetchDetailDialogComponent
+} from './components/catalog-browser-fetch-detail-dialog/catalog-browser-fetch-detail-dialog.component';
 import {CatalogBrowserComponent} from './components/catalog-browser/catalog-browser.component';
 import {ConfirmationDialogComponent} from './components/confirmation-dialog/confirmation-dialog.component';
-import {ContractAgreementTransferDialog} from './components/contract-agreement-transfer-dialog/contract-agreement-transfer-dialog.component';
-import {ContractDefinitionCardsComponent} from './components/contract-definition-cards/contract-definition-cards.component';
-import {ContractDefinitionEditorDialog} from './components/contract-definition-editor-dialog/contract-definition-editor-dialog.component';
-import {ContractDefinitionPageComponent} from './components/contract-definition-page/contract-definition-page.component';
+import {
+  ContractAgreementTransferDialog
+} from './components/contract-agreement-transfer-dialog/contract-agreement-transfer-dialog.component';
+import {
+  ContractDefinitionCardsComponent
+} from './components/contract-definition-cards/contract-definition-cards.component';
+import {
+  ContractDefinitionEditorDialog
+} from './components/contract-definition-editor-dialog/contract-definition-editor-dialog.component';
+import {
+  ContractDefinitionPageComponent
+} from './components/contract-definition-page/contract-definition-page.component';
 import {ContractOfferIconComponent} from './components/contract-offer-icon/contract-offer-icon.component';
 import {ContractOfferListComponent} from './components/contract-offer-list/contract-offer-list.component';
 import {ContractViewerComponent} from './components/contract-viewer/contract-viewer.component';
@@ -69,6 +80,7 @@ import {TransportModeSelectComponent} from './components/transport-mode-select/t
 import {IsActiveFeaturePipe} from './pipes/is-active-feature.pipe';
 import {ReplacePipe} from './pipes/replace.pipe';
 import {SafePipe} from './pipes/safe.pipe';
+
 
 @NgModule({
   imports: [
@@ -118,6 +130,7 @@ import {SafePipe} from './pipes/safe.pipe';
     AssetDetailDialog,
     AssetEditorDialog,
     AssetListComponent,
+    AssetSelectComponent,
     AssetViewerComponent,
     CatalogBrowserComponent,
     CatalogBrowserFetchDetailDialogComponent,
@@ -167,4 +180,5 @@ import {SafePipe} from './pipes/safe.pipe';
     TransferHistoryViewerComponent,
   ],
 })
-export class EdcDemoModule {}
+export class EdcDemoModule {
+}